requirements.txt
google-generativeai>=0.6.0
pillow>=10.1.0
numpy>=1.26.0
pydantic>=2.6.0
typer>=0.12.0
tqdm>=4.66.0
imageio>=2.35.0
imageio-ffmpeg>=0.4.9
pyyaml>=6.0.1
rich>=13.7.0
python-dotenv>=1.0.1
